requirements.txt
Django==4.0.5
crispy-tailwind==0.5.0
django-crispy-forms==1.14.0
django-extra-views==0.14.0
